refactor(containers): split container logs tab
